requirements.txt
certifi==2019.11.28
chardet==3.0.4
cycler==0.10.0
dominate==2.5.1
idna==2.9
joblib==0.14.1
jsonpatch==1.25
jsonpointer==2.0
kiwisolver==1.1.0
matplotlib==3.0.3
numpy==1.18.1
opencv-python==4.2.0.32
pandas==0.25.3
Pillow==7.0.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
pyzmq==19.0.0
requests==2.23.0
scikit-learn==0.22.2.post1
scipy==1.4.1
Shapely==1.7.0
six==1.14.0
sklearn==0.0
torch==1.4.0
torchfile==0.1.0
torchvision==0.5.0
tornado==6.0.4
tqdm==4.46.0
urllib3==1.25.8
visdom==0.1.8.9
websocket-client==0.57.0
